worker/requirements.txt
boto3==1.35.57
ffmpeg-python==0.2.0
PyMySQL==1.1.1

